refactor: 일기 상세 뷰 컨트롤러 메서드 순서 조정 및 주석 추가
